+++ src/main/java/com/purbon/kafka/topology/KafkaTopologyBuilder.java
@@ -5,6 +5,8 @@ import static com.purbon.kafka.topology.TopologyBuilderConfig.REDIS_PORT_CONFIG;
 import static com.purbon.kafka.topology.TopologyBuilderConfig.REDIS_STATE_PROCESSOR_CLASS;
 import static com.purbon.kafka.topology.TopologyBuilderConfig.STATE_PROCESSOR_DEFAULT_CLASS;
 
+import com.purbon.kafka.topology.api.adminclient.TopologyBuilderAdminClient;
+import com.purbon.kafka.topology.api.adminclient.TopologyBuilderAdminClientBuilder;
 import com.purbon.kafka.topology.api.mds.MDSApiClientBuilder;
 import com.purbon.kafka.topology.clusterstate.FileStateProcessor;
 import com.purbon.kafka.topology.clusterstate.RedisStateProcessor;
