+++ src/main/java/com/purbon/kafka/topology/roles/SimpleAclsProvider.java
@@ -1,7 +1,7 @@
 package com.purbon.kafka.topology.roles;
 
 import com.purbon.kafka.topology.AccessControlProvider;
-import com.purbon.kafka.topology.TopologyBuilderAdminClient;
+import com.purbon.kafka.topology.api.adminclient.TopologyBuilderAdminClient;
 import java.io.IOException;
 import java.util.HashMap;
 import java.util.List;
